Implement prototype of game
Adds grid board for playing the game. Starts with player using X. Game checks when any of the rows, columns or diagonals are filled with a player's X or O.

Has reset button to clear the board state.

Two divs to send message when player wins the game, and when player tries to choose an already filled square.